requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0 